feat: respect stories paths defined in main.js
- On the way I also refactored the entire folder in src/util to allow for easier module mocking in tests
