+++ src/util/getCliOptions.ts
@@ -1,6 +1,4 @@
-import { join, resolve } from 'path';
-import { serverRequire, StorybookConfig } from '@storybook/core-common';
-import { getParsedCliOptions } from './helpers';
+import { getParsedCliOptions } from './getParsedCliOptions';
 
 type CliOptions = {
   runnerOptions: {
@@ -18,8 +16,6 @@ export const defaultRunnerOptions: CliOptions['runnerOptions'] = {
   configDir: '.storybook',
   storiesJson: false,
 };
-
-let storybookMainConfig: StorybookConfig;
 
 export const getCliOptions = () => {
   const allOptions = getParsedCliOptions();
@@ -45,19 +41,4 @@ export const getCliOptions = () => {
 
     return acc;
   }, defaultOptions);
-};
-
-export const getStorybookMain = (configDir: string) => {
-  if (storybookMainConfig) {
-    return storybookMainConfig;
-  }
-
-  storybookMainConfig = serverRequire(join(resolve(configDir), 'main'));
-  if (!storybookMainConfig) {
-    throw new Error(
-      `Could not load main.js in ${configDir}. Is the config directory correct? You can change it by using --config-dir <path-to-dir>`
-    );
-  }
-
-  return storybookMainConfig;
 };
