+++ src/util/getStorybookMetadata.test.ts
@@ -1,0 +1,68 @@
+import * as storybookMain from './getStorybookMain';
+
+import { getStorybookMetadata } from './getStorybookMetadata';
+
+describe.only('getStorybookMetadata', () => {
+  afterAll(() => {
+    process.env.STORYBOOK_CONFIG_DIR = undefined;
+  });
+
+  it('should return configDir coming from environment variable', () => {
+    const mockedMain = {
+      stories: [],
+    };
+
+    jest.spyOn(storybookMain, 'getStorybookMain').mockReturnValueOnce(mockedMain);
+    process.env.STORYBOOK_CONFIG_DIR = '.storybook';
+    const { configDir } = getStorybookMetadata();
+    expect(configDir).toEqual(process.env.STORYBOOK_CONFIG_DIR);
+  });
+
+  it('should return storiesPath with default glob from CSF3 style config', () => {
+    const mockedMain = {
+      stories: [
+        {
+          directory: '../stories/basic',
+          titlePrefix: 'Example',
+        },
+      ],
+    };
+
+    jest.spyOn(storybookMain, 'getStorybookMain').mockReturnValueOnce(mockedMain);
+    process.env.STORYBOOK_CONFIG_DIR = '.storybook';
+    const { storiesPaths } = getStorybookMetadata();
+    expect(storiesPaths).toMatchInlineSnapshot(
+      `"<rootDir>/stories/basic/**/*.stories.@(mdx|tsx|ts|jsx|js)"`
+    );
+  });
+
+  it('should return storiesPath with glob defined in CSF2 style config', () => {
+    const mockedMain = {
+      stories: ['../**/stories/*.stories.@(js|ts)'],
+    };
+
+    jest.spyOn(storybookMain, 'getStorybookMain').mockReturnValueOnce(mockedMain);
+    process.env.STORYBOOK_CONFIG_DIR = '.storybook';
+    const { storiesPaths } = getStorybookMetadata();
+    expect(storiesPaths).toMatchInlineSnapshot(`"<rootDir>/**/stories/*.stories.@(js|ts)"`);
+  });
+
+  it('should return storiesPath from mixed CSF2 and CSF3 style config', () => {
+    const mockedMain = {
+      stories: [
+        {
+          directory: '../stories/basic',
+          titlePrefix: 'Example',
+        },
+        '../stories/complex/*.stories.@(js|ts)',
+      ],
+    };
+
+    jest.spyOn(storybookMain, 'getStorybookMain').mockReturnValueOnce(mockedMain);
+    process.env.STORYBOOK_CONFIG_DIR = '.storybook';
+    const { storiesPaths } = getStorybookMetadata();
+    expect(storiesPaths).toMatchInlineSnapshot(
+      `"<rootDir>/stories/basic/**/*.stories.@(mdx|tsx|ts|jsx|js);<rootDir>/stories/complex/*.stories.@(js|ts)"`
+    );
+  });
+});
